Barriers updated
I fixed the barriers in the house so you can no longer run on the wall. The barrier on the guardian island has been fix as well. You can no longer run of the side into the rocks. The guardian island has been updated with trees below the shrine to prevent the glitch where you push up to continuously change scenes. There is a new glitch now in that when you return to the island you spawn on the tree and can no longer move.